Add mouse support to stdlib.p8.
Pico supports mouse control, but its a bit hidden and not recomended.
This change adds a block (surrounded by -- @{ -- @}) that adds a
make_mouse_pointer object that will change colors when the user clicks
one of the mouse buttons.  It also prints out the x, y coordinates (in
screen space) of the cursor.  It also adds an add_gobj function that
returns the newly created object as well as adding it to the g_objs,
which is handy for creating something and adding it right away to the
gobjs.