Uses the add_gobjs function to save a few tokens.
The built in add function doesn't return the object being added, so if
you want to both add a new thing to a list and save a global (or local)
name for it, you have to do it in two steps.  This saves a few tokens by
doing it in one call.